refactor: share download engine
